+++ aqueduct_test/test/auth_harness_test.dart
@@ -150,4 +150,4 @@ class HarnessSubclass extends TestHarness<Channel> with TestHarnessManagedAuthMi
 
 class User extends ManagedObject<_User> implements _User, ManagedAuthResourceOwner<_User> {}
 
-class _User extends ManagedAuthenticatable {}
+class _User extends ResourceOwnerTableDefinition {}