+++ aqueduct/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 3.0.0
 
 - `package:aqueduct/test` moved to `package:aqueduct_test/aqueduct_test`, which is a separate dependency than `aqueduct`.
+- Renames methods in `AuthDelegate` to provide consistency.
+- Adds `AuthDelegate.addClient` and `AuthServer.addClient`.
 - Adds `ManagedContext.transaction` to enable queries to be run in a database transaction. 
 - Removes `ManagedContext.defaultContext`; context usage must be explicit.
 - Adds 'Scope' annotation to add granular scoping to `ResourceController` methods.
@@ -26,6 +28,7 @@
     - Adds concrete implementations of `AuthorizationParser<T>`, `AuthorizationBearerParser` and `AuthorizationBasicParser`.
 - Removes `AuthValidator.fromBearerToken` and `AuthValidator.fromBasicCredentials` and replaces with `AuthValidator.validate<T>`.    
 - Renames the following:
+    - `Authorization.resourceOwnerIdentifier` -> `Authorization.ownerID`
     - `Request.innerRequest` -> `Request.raw`
     - `whereContainsString` -> `whereContains`
     - `AuthStorage` -> `AuthServerDelegate`